[JBTM-2905] moving JCAServerTransactionHeaderReader to package available during runtime
