+++ ArjunaCore/arjuna/classes/com/arjuna/ats/arjuna/tools/osb/mbean/ObjStoreBrowser.java
@@ -95,7 +95,7 @@ public class ObjStoreBrowser implements ObjStoreBrowserMBean {
                     "com.arjuna.ats.internal.jta.tools.osb.mbean.jts.JCAServerTransactionWrapper",
                     "com.arjuna.ats.internal.jta.tools.osb.mbean.jts.JTSActionBean",
                     SUBORDINATE_ATI_TYPE,
-                    "com.hp.mwtests.ts.jta.jts.tools.JCAServerTransactionHeaderReader"
+                    "com.arjuna.ats.internal.jta.tools.osb.mbean.jts.JCAServerTransactionHeaderReader"
             ),
             new OSBTypeHandler(
                     true,
