+++ ArjunaJTS/jtax/tests/classes/com/hp/mwtests/ts/jta/jts/tools/NewTypeTest.java
@@ -32,6 +32,7 @@ import com.arjuna.ats.arjuna.common.Uid;
 import com.arjuna.ats.arjuna.tools.osb.mbean.OSBTypeHandler;
 import com.arjuna.ats.arjuna.tools.osb.mbean.ObjStoreBrowser;
 import com.arjuna.ats.arjuna.tools.osb.mbean.UidWrapper;
+import com.arjuna.ats.internal.jta.tools.osb.mbean.jts.JCAServerTransactionHeaderReader;
 /**
  * An example of how to instrument new record types.
  *
